backend/requirements.txt
# ──────────────────────────────────────────────
# QuietMonitor – Python Backend Dependencies
# ──────────────────────────────────────────────
# Version ranges keep this compatible with Python 3.10–3.14.
# If pydantic-core fails to build on your Python version,
# run:  pip install --pre pydantic pydantic-core pydantic-settings

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
alembic>=1.13.1
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-multipart>=0.0.9
